# internal/e2bapi/openapi.json
{
  "components": {
...
    "responses": {
      "400": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Bad request"
      },
      "401": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Authentication error"
      },
...
      "500": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Server error"
      }
    },
    "schemas": {
...
      "CPUCount": {
        "description": "CPU cores for the sandbox",
        "format": "int32",
        "minimum": 1,
        "type": "integer"
      },
...
      "DiskSizeMB": {
        "description": "Disk size for the sandbox in MiB",
        "format": "int32",
        "minimum": 0,
        "type": "integer"
      },
...
      "EnvdVersion": {
        "description": "Version of the envd running in the sandbox",
        "type": "string"
      },
      "Error": {
        "properties": {
          "code": {
            "description": "Error code",
            "format": "int32",
            "type": "integer"
          },
          "message": {
            "description": "Error",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "MemoryMB": {
        "description": "Memory for the sandbox in MiB",
        "format": "int32",
        "minimum": 128,
        "type": "integer"
      },
...
      "TeamUser": {
        "properties": {
          "email": {
            "description": "Email of the user",
            "type": "string"
          },
          "id": {
            "description": "Identifier of the user",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "Template": {
        "properties": {
          "aliases": {
            "deprecated": true,
            "description": "Aliases of the template",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "buildCount": {
            "description": "Number of times the template was built",
            "format": "int32",
            "type": "integer"
          },
          "buildID": {
            "description": "Identifier of the last successful build for given template",
            "type": "string"
          },
          "buildStatus": {
            "$ref": "#/components/schemas/TemplateBuildStatus"
          },
          "cpuCount": {
            "$ref": "#/components/schemas/CPUCount"
          },
          "createdAt": {
            "description": "Time when the template was created",
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TeamUser"
              }
            ],
            "nullable": true
          },
          "diskSizeMB": {
            "$ref": "#/components/schemas/DiskSizeMB"
          },
          "envdVersion": {
            "$ref": "#/components/schemas/EnvdVersion"
          },
          "lastSpawnedAt": {
            "description": "Time when the template was last used",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "memoryMB": {
            "$ref": "#/components/schemas/MemoryMB"
          },
          "names": {
            "description": "Names of the template (namespace/alias format when namespaced)",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "public": {
            "description": "Whether the template is public or only accessible by the team",
            "type": "boolean"
          },
          "spawnCount": {
            "description": "Number of times the template was used",
            "format": "int64",
            "type": "integer"
          },
          "templateID": {
            "description": "Identifier of the template",
            "type": "string"
          },
          "updatedAt": {
            "description": "Time when the template was last updated",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "templateID",
          "buildID",
          "cpuCount",
          "memoryMB",
          "diskSizeMB",
          "public",
          "createdAt",
          "updatedAt",
          "createdBy",
          "lastSpawnedAt",
          "spawnCount",
          "buildCount",
          "envdVersion",
          "aliases",
          "names",
          "buildStatus"
        ]
      },
...
      "TemplateBuildRequest": {
        "properties": {
          "alias": {
            "description": "Alias of the template",
            "type": "string"
          },
          "cpuCount": {
            "$ref": "#/components/schemas/CPUCount"
          },
          "dockerfile": {
            "description": "Dockerfile for the template",
            "type": "string"
          },
          "memoryMB": {
            "$ref": "#/components/schemas/MemoryMB"
          },
          "readyCmd": {
            "description": "Ready check command to execute in the template after the build",
            "type": "string"
          },
          "startCmd": {
            "description": "Start command to execute in the template after the build",
            "type": "string"
          },
          "teamID": {
            "description": "Identifier of the team",
            "type": "string"
          }
        },
        "required": [
          "dockerfile"
        ]
      },
      "TemplateBuildRequestV2": {
        "properties": {
          "alias": {
            "description": "Alias of the template",
            "type": "string"
          },
          "cpuCount": {
            "$ref": "#/components/schemas/CPUCount"
          },
          "memoryMB": {
            "$ref": "#/components/schemas/MemoryMB"
          },
          "teamID": {
            "deprecated": true,
            "description": "Identifier of the team",
            "type": "string"
          }
        },
        "required": [
          "alias"
        ]
      },
      "TemplateBuildRequestV3": {
        "properties": {
          "alias": {
            "deprecated": true,
            "description": "Alias of the template. Deprecated, use name instead.",
            "type": "string"
          },
          "cpuCount": {
            "$ref": "#/components/schemas/CPUCount"
          },
          "memoryMB": {
            "$ref": "#/components/schemas/MemoryMB"
          },
          "name": {
            "description": "Name of the template. Can include a tag with colon separator (e.g. \"my-template\" or \"my-template:v1\"). If tag is included, it will be treated as if the tag was provided in the tags array.",
            "type": "string"
          },
          "tags": {
            "description": "Tags to assign to the template build",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "teamID": {
            "deprecated": true,
            "description": "Identifier of the team",
            "type": "string"
          }
        }
      },
...
      "TemplateBuildStatus": {
        "description": "Status of the template build",
        "enum": [
          "building",
          "waiting",
          "ready",
          "error"
        ],
        "type": "string"
      },
      "TemplateLegacy": {
        "properties": {
          "aliases": {
            "description": "Aliases of the template",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "buildCount": {
            "description": "Number of times the template was built",
            "format": "int32",
            "type": "integer"
          },
          "buildID": {
            "description": "Identifier of the last successful build for given template",
            "type": "string"
          },
          "cpuCount": {
            "$ref": "#/components/schemas/CPUCount"
          },
          "createdAt": {
            "description": "Time when the template was created",
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TeamUser"
              }
            ],
            "nullable": true
          },
          "diskSizeMB": {
            "$ref": "#/components/schemas/DiskSizeMB"
          },
          "envdVersion": {
            "$ref": "#/components/schemas/EnvdVersion"
          },
          "lastSpawnedAt": {
            "description": "Time when the template was last used",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "memoryMB": {
            "$ref": "#/components/schemas/MemoryMB"
          },
          "public": {
            "description": "Whether the template is public or only accessible by the team",
            "type": "boolean"
          },
          "spawnCount": {
            "description": "Number of times the template was used",
            "format": "int64",
            "type": "integer"
          },
          "templateID": {
            "description": "Identifier of the template",
            "type": "string"
          },
          "updatedAt": {
            "description": "Time when the template was last updated",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "templateID",
          "buildID",
          "cpuCount",
          "memoryMB",
          "diskSizeMB",
          "public",
          "createdAt",
          "updatedAt",
          "createdBy",
          "lastSpawnedAt",
          "spawnCount",
          "buildCount",
          "envdVersion",
          "aliases"
        ]
      },
      "TemplateRequestResponseV3": {
        "properties": {
          "aliases": {
            "deprecated": true,
            "description": "Aliases of the template",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "buildID": {
            "description": "Identifier of the last successful build for given template",
            "type": "string"
          },
          "names": {
            "description": "Names of the template",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "public": {
            "description": "Whether the template is public or only accessible by the team",
            "type": "boolean"
          },
          "tags": {
            "description": "Tags assigned to the template build",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "templateID": {
            "description": "Identifier of the template",
            "type": "string"
          }
        },
        "required": [
          "templateID",
          "buildID",
          "public",
          "aliases",
          "names",
          "tags"
        ]
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "bearerFormat": "access_token",
        "scheme": "bearer",
        "type": "http"
      },
      "AdminTokenAuth": {
        "in": "header",
        "name": "X-Admin-Token",
        "type": "apiKey"
      },
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "Supabase1TokenAuth": {
        "in": "header",
        "name": "X-Supabase-Token",
        "type": "apiKey"
      },
      "Supabase2TeamAuth": {
        "in": "header",
        "name": "X-Supabase-Team",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "E2B API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/templates": {
      "get": {
        "description": "List all templates",
        "operationId": "GetTemplates",
        "parameters": [
          {
            "in": "query",
            "name": "teamID",
            "schema": {
              "description": "Identifier of the team",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Template"
                      }
                    ]
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successfully returned all templates"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "AccessTokenAuth": []
          },
          {
            "Supabase1TokenAuth": [],
            "Supabase2TeamAuth": []
          }
        ],
        "tags": [
          "templates"
        ]
      },
      "post": {
        "deprecated": true,
        "description": "Create a new template",
        "operationId": "PostTemplates",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateLegacy"
                }
              }
            },
            "description": "The build was accepted"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          },
          {
            "Supabase1TokenAuth": [],
            "Supabase2TeamAuth": []
          }
        ],
        "tags": [
          "templates"
        ]
      }
    },
...
    "/v2/templates": {
      "post": {
        "deprecated": true,
        "description": "Create a new template",
        "operationId": "PostV2Templates",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateBuildRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateLegacy"
                }
              }
            },
            "description": "The build was requested successfully"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "Supabase1TokenAuth": [],
            "Supabase2TeamAuth": []
          }
        ],
        "tags": [
          "templates"
        ]
      }
    },
...
    "/v3/templates": {
      "post": {
        "description": "Create a new template",
        "operationId": "PostV3Templates",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateBuildRequestV3"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateRequestResponseV3"
                }
              }
            },
            "description": "The build was requested successfully"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "Supabase1TokenAuth": [],
            "Supabase2TeamAuth": []
          }
        ],
        "tags": [
          "templates"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.e2b.app"
    }
  ],
...
}